🚚 Move `IntegerJavaSample` to `org.kotools.types.number` (#939)
